Rename `LineLocation` to `SourceLocation`
It represents a location in the source code, so I think the new name is
more fitting. Talking about line numbers there is misleading anyway, as
there is no such thing as line-based formatting in the source form right
now, and how the debugger shows it is basically an implementation
detail.