Remove `extern "C"` from FFI functions
I don't think it's actually needed, as those functions are only used to
interact with JavaScript over the module boundary. Presumably, a stable
interface for this is already specified (if not, how would the
JavaScript side work?).

Also, `extern "C"` is limiting, as WebAssembly supports multi-value
returns, while the C ABI doesn't.