Update name of `SyntaxBuilder` function
What was, for a brief time, called a "reference" now uses the more
standard name of "identifier". This commit updates the respective
function of `SyntaxBuilder` to reflect that.

The rename also switches away from the older "one character" naming
convention. This should be a bit more readable, and hopefully (thanks to
IDE support) not much harder to write.