Simplify return value
With the impending removal of the legacy stack, the question of whether
a stack is empty, and if that's a problem, is less clear-cut. Removing
the return value is a convenient way to avoid having to add non-trivial
logic to the function.

It also seems that all callers are pretty sure, that this wasn't going
to happen anyway.

Maybe the return value should be replaced by an assertion within the
function, but as of now, I'm not sure what that assertion would be.