Put bindings on initial stack frame
This makes the initial stack frame look like a regular stack frame,
except that the return address is substituted for a start marker. This
doesn't make a difference right now (except making the stack slightly
larger), but it's going to help with a necessary rewrite of tail call
elimination. That should be coming up in the next commit.