Fix tail call elimination removing start marker
The call from the initial stack frame is always a tail call. Tail call
elimination, as previously implemented, would always remove the start
marker on that call, which meant that the start marker didn't work as
documented.

This didn't make a difference, because it was being masked by another
bug. I'm in the process of fixing that.