Remove placeholder pass
After the simplification I did to `e_bytecode` to prepare for extracting
this pass, the pass is no longer necessary in the first place.
`e_bytecode` is now not overly complicated, so it's not a pressing issue
anymore. And doing replacements in a separate pass would require so many
more return value from `e_bytecode`, it wouldn't be worth it on balance.