Don't reset process on instruction update
This allows loading new instruction directly into the running process.
I've confirmed that this works for very simple cases, but I'm sure it
won't for any more complex ones. This will lead to instability.

It's fine for now. The priority from here is to identify the sources of
these problems, and address them one by one.