Stop on breakpoint before evaling next instruction
This seems to have no adverse effects, and it makes sense. It's a change
I've been wanting to make for a while, and it seems that in the
meantime, it's actually become easy to do so.

The reason I'm doing it right now, is that it fixes a weird interaction
with the handling of other effects, which caused a confusing list of
active functions in some cases.