Prepare for not re-resolving host functions
This makes the output in the debugger worse, for the time being. Which
is unfortunate, but the debugger shouldn't be relying on this code
anyway. (And in fact, I think this code shouldn't exist.)

The debugger shuold match on the fragment payload directly and then do
proper syntax highlighting based on what it sees. Once it does that,
this `Display` implementation can be removed.