Add `Evaluator::evaluate_instruction`
This is the start of a process that changes how effects affect
`next_instruction`. The goal is to only advance to the next instruction,
if no effect was triggered.

This will bring breakpoints triggered by `brk` in line with how user-set
breakpoints already work. Currently, a lot seems to be broken in that
regard.