Prepare for double-sided hashing
I think I'm going to need to hash fragments from both sides. Right-to-
left, as right now, but also left-to-right in addition. This is going to
be necessary to determine which code has been updated.

By separating compilation and hashing, like I did here, I'm paving the
way for that to happen.