Prepare for special handling of recursive calls
I'm working on refering to called functions by fragment ID as opposed to
by name. This is something I've been wanting to do for a while, but it
has now become a higher priority, as it's going to help with the
improved handling of code updates that I'm planning.

A recursive call can't be done by ID. To hash the call, we need the hash
of the called function, but for that we need the hash of the call. So we
need special handling for those cases, and this is what this commit
prepares for.