Add partial support for type inference
So far, only inference for bindings is implemented. Since that can't
provide any information, this is not useful.

The point of adding this partial implementation, is to now have the
option of passing this information to the debugger and displaying it
there. This will make testing easier, as I implement more of type
inference.