Make better use of type annotations in inference
This is a significant improvement to type inference. It makes better use
of type annotations by using the `unify` method to unify inferred and
explicit signatures, instead of trying to do a similar thing with
purpose-built but redundant code.

Since this removes said redundant code, this is also a simplification.