Improve type inference of bindings
All I'm doing here is remove a struct field that has become redundant.
It seems that this struct field was somehow now always set correctly,
preventing types from being unified.

I don't understand that bug in any more detail than that, but the
inference its removal enables it legit. The new inference matched the
annotations that I removed.

Given that I wanted to remove this field anyway, because it's weird and
redundant, I'm taking the gift, instead of digging into a bug that no
longer exists.